+++ sdlm/models/mixins/modeling_mixin.py
@@ -45,9 +45,7 @@ class DiffusionModelMixin:
         inputs_word_embeds = self.get_input_embeddings()(input_ids)
         if not self.config.disable_timestep_embed:
             timesteps = torch.where(span_mask, timesteps, torch.zeros_like(timesteps))
-            timesteps_embed = self.timestep_embed(
-                timesteps.unsqueeze(-1).to(self.dtype)
-            )
+            timesteps_embed = self.timestep_embed(timesteps.unsqueeze(-1).float())
             inputs_embeds = inputs_embeds + timesteps_embed
         # For the unmasked tokens, we only compute their original word embeddings.
         # Note that this also sets the self-conditioned inputs which we are conditioning on
